Restructured ToggleBtn functionality so that it's a child component
