+++ src/client/components/display/OpenAllBtn.jsx
@@ -1,8 +1,13 @@
 import React from 'react';
+import ReqResCtrl from '../ReqResCtrl';
 
 const OpenAllBtn = () => {
   return (
-    <button type='button'>OpenAllBtn</button>
+    <button type='button' onClick={
+      (e) => { 
+        ReqResCtrl.openAllEndPoints(e)
+      }
+    }>OpenAllBtn</button>
   );
 };
