+++ src/client/components/display/ReqRes.jsx
@@ -1,12 +1,9 @@
 import React, { Component } from "react";
 import { connect } from 'react-redux';
-
-import * as store from '../../store';
-
 import Request from './Request.jsx';
 import Response from './Response.jsx';
+import ToggleBtn from './ToggleBtn.jsx';
 import ReqResCtrl from '../ReqResCtrl';
-
 import * as actions from '../../actions/actions';
 
 const mapStateToProps = store => ({
@@ -20,40 +17,6 @@ const mapDispatchToProps = dispatch => ({
 class ReqRes extends Component {
   constructor(props) {
     super(props);
-    this.state = {
-      isToggled: true,
-      prevContentBody: [],
-    };
-    this.handleToggleClick = this.handleToggleClick.bind(this);
-}
-
-  componentDidMount() {
-    this.setState({
-      abortController : new AbortController(),
-    })
-  }
-
-  handleToggleClick(e, abortCtrl) {
-    console.log('isToggled', this.state.isToggled);
-    if (this.state.isToggled) {
-      ReqResCtrl.toggleOpenEndPoint(e, abortCtrl);
-    } else {
-      console.log('closeEndpoint', e.target);
-      const reqResComponentID = e.target.id;
-      const gotState = store.default.getState();
-      const reqResArr = gotState.business.reqResArray;
-      const targetReqResObj = reqResArr.find(obj => obj.id == reqResComponentID); 
-
-      targetReqResObj.connection = 'closed';
-      this.state.abortController.abort();
-      this.setState({
-        abortController : new AbortController(),
-      })
-    }
-
-    this.setState(prevState => ({
-        isToggled: !prevState.isToggled
-    }));
   }
 
   render() {
@@ -71,14 +34,7 @@ class ReqRes extends Component {
         {this.props.content.timeReceived}
         {this.props.content.connectionType}
         {contentBody}
-        
-        <button id={this.props.content.id}  onClick={
-          (e) => { 
-            this.handleToggleClick(e, this.state.abortController)
-          }
-        }>
-          {this.state.isToggled ? 'OPEN CONNECTION' : 'CLOSE CONNECTION'}
-        </button>
+        <ToggleBtn reqResState={this.props}/>
       </div>
     )
   }
